+++ src/chanjo2/models/pydantic_models.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from typing import Any, Dict, List, Optional, Union
 
 import validators
-from pydantic import BaseModel, Field, field_validator, model_validator
+from pydantic import BaseModel, Field, FilePath, field_validator, model_validator
 from pydantic_settings import SettingsConfigDict
 from starlette.datastructures import FormData
 
@@ -217,6 +217,19 @@ class SampleGeneIntervalQuery(BaseModel):
 
 
 ## Coverage and  overview report - related models ###
+
+
+class CoverageSummaryQuerySample(BaseModel):
+    name: str
+    coverage_file_path: str
+
+
+class CoverageSummaryQuery(BaseModel):
+    build: Builds
+    samples: List[CoverageSummaryQuerySample]
+    hgnc_gene_ids: List[int]
+    coverage_threshold: int
+    interval_type: IntervalType
 
 
 class ReportQuerySample(BaseModel):